🌌 move token to interseptor, add redirects, remove unnecessary components
