+++ src/api/requests/getOrders.ts
@@ -1,15 +1,6 @@
 import { instance } from '../instance'
 
-export interface GetOrdersParams {
-   token: string
-}
+type getOrdersRequestConfig = RequestConfig
 
-type getOrdersRequestConfig = RequestConfig<GetOrdersParams>
-
-export const getOrders = ({ params, config }: getOrdersRequestConfig) =>
-   instance.get<GetOrdersResponse>(`/cinema/orders`, {
-      headers: {
-         Authorization: `Bearer ${params.token}`
-      },
-      ...config
-   })
+export const getOrders = ({ config }: getOrdersRequestConfig) =>
+   instance.get<GetOrdersResponse>(`/cinema/orders`, config)
