+++ src/api/requests/getSession.ts
@@ -1,15 +1,6 @@
 import { instance } from '../instance'
 
-interface GetSessionParams {
-   token: string
-}
+type getSessionRequestConfig = RequestConfig
 
-type getSessionRequestConfig = RequestConfig<GetSessionParams>
-
-export const getSession = ({ params, config }: getSessionRequestConfig) =>
-   instance.get<GetSessionResponse>(`/users/session`, {
-      headers: {
-         Authorization: `Bearer ${params.token}`
-      },
-      ...config
-   })
+export const getSession = ({ config }: getSessionRequestConfig) =>
+   instance.get<GetSessionResponse>(`/users/session`, config)
