--- src/app/[filmId]/(components)/LayoutOrderProvider/LayoutOrderProvider.tsx
@@ -1,9 +1,0 @@
-'use client'
-
-import { ReactNode } from 'react'
-
-import { OrderContextProvider } from '@/contexts/orderContext/orderContextProvider'
-
-export  const LayoutOrderProvider = ({ children }: { children: ReactNode }) => {
-   return <OrderContextProvider>{children}</OrderContextProvider>
-}
