+++ src/app/[filmId]/layout.tsx
@@ -2,18 +2,18 @@ import { ReactNode } from 'react'
 
 import { getFilm } from '@/api/requests/getFilm'
 
-import { LayoutOrderProvider } from './(components)/LayoutOrderProvider/LayoutOrderProvider'
+import { OrderContextProvider } from '@/contexts/orderContext/orderContextProvider'
 
 export async function generateMetadata({ params }: { params: { filmId: string } }) {
    const getFilmResponse = await getFilm({ params: { id: params.filmId } })
-   
+
    return {
       title: `${getFilmResponse.data.film.name} - ШИФТ CINEMA`
    }
 }
 
 const OrderLayout = ({ children }: { children: ReactNode }) => {
-   return <LayoutOrderProvider>{children}</LayoutOrderProvider>
+   return <OrderContextProvider>{children}</OrderContextProvider>
 }
 
 export default OrderLayout
