+++ src/app/orders/(components)/OrdersTabs/OrdersTabs.module.css
@@ -17,6 +17,7 @@
 .card {
    display: flex;
    flex-direction: column;
+   justify-content: space-between;
    gap: 19px;
    border-radius: 16px;
 }
