+++ src/helpers/isSSR.ts
@@ -1,0 +1,1 @@
+export const isSSR = typeof window === "undefined"
